Clean up 32-bit ass-u-mptions for x64 and arm64.

+++ src/dutil/inc/xmlutil.h
@@ -53,7 +53,7 @@ HRESULT DAPI XmlLoadDocumentFromFile(
     );
 HRESULT DAPI XmlLoadDocumentFromBuffer(
     __in_bcount(cbSource) const BYTE* pbSource,
-    __in DWORD cbSource,
+    __in SIZE_T cbSource,
     __out IXMLDOMDocument** ppixdDocument
     );
 HRESULT DAPI XmlLoadDocumentFromFileEx(
